Phase 1: Extract config, scoring, translation modules from app.py
Structural decomposition — no behaviour changes. app.py drops from
4,034 to 3,300 lines (734 lines extracted into focused modules).

New modules:
- src/config.py: LANGUAGE_CONFIG, voice mappings, DEFAULT_SETTINGS,
  load_settings/save_settings (now accept session_state/db_module params)
- src/scoring/comparison.py: levenshtein_distance, edit operations,
  compare_phonemes (positional + edit distance)
- src/scoring/phonemes.py: get_espeak_path, get_phonemes, get_ipa,
  normalize_for_phoneme_scoring, format_ipa, get_ipa_from_espeak
- src/translation.py: get_translation_provider, validate_translation_api_key,
  get_translation_from_llm, enrich_material_file

app.py now imports from these modules. Thin wrappers remain for
load_settings/save_settings to pass st.session_state and app_mysql.

Other changes:
- Tests updated to import from real modules (no more duplicated code)
- 3 new config tests (28 total, all passing)
- clean_restart.py: worktree-compatible (finds workspace via src/ not
  venv/, uses active venv's streamlit)
- .gitignore: exclude src/config.py from autotools config.* pattern
- REFACTOR_PLAN.md: Phase 6 additions (provenance, vocab extensibility)

Verified: app starts cleanly, all 28 tests pass, clean_restart.py works
in worktree.
