+++ src/app.py
@@ -8,144 +8,29 @@ with real-time feedback using speech recognition and phonetic analysis.
 Run with: streamlit run app.py
 """
 
-# VERSION MARKER - Update this when releasing new version
-__version__ = "7.1.3"
-__app_name__ = "Pronunciation Trainer"
-__author__ = "Matthew Fairtlough & Contributors"
-__license__ = "GPL-3.0"
+# Configuration, constants, and language definitions — now in config.py
+from config import (
+    __version__, __app_name__, __author__, __license__,
+    LANGUAGE_CONFIG, VOICE_LOCALE_NORMALIZATION, GOOGLE_CLOUD_VOICES,
+    DEFAULT_SETTINGS, get_language_code, get_language_for_provider,
+)
 
-# Language configuration
-LANGUAGE_CONFIG = {
-    "Portuguese": {
-        "code": "pt",
-        "display_name": "Portuguese Pronunciation Trainer",
-        "voices": {
-            "google_cloud": ["pt-br", "pt"],
-            "gtts": ["pt-br", "pt"],
-            "espeak": ["pt-br", "pt"]
-        }
-    },
-    "Dutch": {
-        "code": "nl",
-        "display_name": "Dutch/Flemish Pronunciation Trainer",
-        "voices": {
-            "google_cloud": ["nl", "nl-be"],
-            "gtts": ["nl"],
-            "espeak": ["nl"]
-        }
-    },
-    "French": {
-        "code": "fr",
-        "display_name": "French Pronunciation Trainer",
-        "voices": {
-            "google_cloud": ["fr", "fr-fr"],
-            "gtts": ["fr"],
-            "espeak": ["fr-fr"]
-        }
-    },
-    "German": {
-        "code": "de",
-        "display_name": "German Pronunciation Trainer",
-        "voices": {
-            "google_cloud": ["de", "de-de"],
-            "gtts": ["de"],
-            "espeak": ["de"]
-        }
-    },
-    "Italian": {
-        "code": "it",
-        "display_name": "Italian Pronunciation Trainer",
-        "voices": {
-            "google_cloud": ["it", "it-it"],
-            "gtts": ["it"],
-            "espeak": ["it"]
-        }
-    },
-    "Spanish": {
-        "code": "es",
-        "display_name": "Spanish Pronunciation Trainer",
-        "voices": {
-            "google_cloud": ["es", "es-es"],
-            "gtts": ["es"],
-            "espeak": ["es"]
-        }
-    }
-}
+# Scoring and phoneme modules
+from scoring.comparison import (
+    levenshtein_distance, get_edit_operations,
+    compare_phonemes_positional, compare_phonemes_edit_distance,
+    compare_phonemes,
+)
+from scoring.phonemes import (
+    get_espeak_path, get_phonemes, normalize_for_phoneme_scoring,
+    get_ipa, get_ipa_from_espeak, format_ipa,
+)
 
-# Voice locale normalization: lowercase codes → BCP 47 format
-VOICE_LOCALE_NORMALIZATION = {
-    'pt-br': 'pt-BR',
-    'pt': 'pt-PT',
-    'fr': 'fr-FR',
-    'fr-fr': 'fr-FR',
-    'nl': 'nl-NL',
-    'nl-be': 'nl-BE',
-    'de': 'de-DE',
-    'de-de': 'de-DE',
-    'it': 'it-IT',
-    'it-it': 'it-IT',
-    'es': 'es-ES',
-    'es-es': 'es-ES'
-}
-
-# Google Cloud TTS voice names per locale
-GOOGLE_CLOUD_VOICES = {
-    "pt-BR": "pt-BR-Standard-A",  # Female Brazilian Portuguese
-    "pt-PT": "pt-PT-Standard-A",  # Female European Portuguese
-    "fr-FR": "fr-FR-Standard-A",  # Female French
-    "nl-NL": "nl-NL-Standard-A",  # Female Dutch
-    "nl-BE": "nl-BE-Standard-A",  # Female Flemish
-    "de-DE": "de-DE-Standard-A",  # Female German
-    "it-IT": "it-IT-Standard-A",  # Female Italian
-    "es-ES": "es-ES-Standard-A",  # Female Spanish
-}
-
-# Version History:
-# 1.2.1 (2025-11-13):
-#   - Documentation: Update all docs to be language-agnostic (Miolingo branding)
-#   - Update primary URL to miolingo.io with backup streamlit.app URL
-#   - Emphasize multi-language support throughout documentation
-# 1.2.0 (2025-11-13):
-#   - Add built-in language materials library browser (French: 200 phrases + 428 words, Portuguese: 83 phrases + 172 words)
-#   - New tabbed interface: "Built-in Library" + "Upload File" for better UX
-#   - File metadata preview (item count, translations, IPA) before loading
-#   - Track material source in session state
-#   - Add app_language_materials.py module for materials management
-# 1.1.3 (2025-11-13):
-#   - Remove separator line between audio and recording for better mobile spacing
-#   - Dynamic language name in recording instructions (Portuguese/French/Dutch)
-# 1.1.2 (2025-11-13):
-#   - Further mobile UX improvements: smaller phrase heading (h4), info box moved below recording widget
-# 1.1.1 (2025-11-13):
-#   - Fix eSpeak TTS auto-play bug (use --stdout to capture audio bytes)
-#   - Fix mobile UX: smaller heading (h3), emoji inline, translation above phrase
-#   - Add French language materials infrastructure (phrases A-D levels)
-# 1.0.0 (2025-11-11):
-#   - Add multi-language support (Portuguese, French, Dutch, Flemish)
-#   - Language-specific session tracking
-#   - Dynamic app title based on selected language
-#   - Language selector in sidebar
-# 0.9.3 (2025-11-11):
-#   - Improve audio trimming with 200ms padding (prevents speech artifacts)
-#   - Show trimmed audio in results (what was actually recognized)
-#   - Use IPA instead of eIPA for normalized phonemes display (user-friendly)
-#   - Fix Edit button UX: grayed out in edit mode, clear "Return to Guided Mode" button
-#   - British spelling: "Normalised" in UI
-# 0.9.2 (2025-11-10):
-#   - Add TTS engine selection (Google TTS vs eSpeak)
-#   - Fix speed/pitch settings now work (with eSpeak)
-#   - Add slow mode for Google TTS (simple on/off)
-#   - Add documentation links in sidebar
-#   - Reorganize documentation (USER_GUIDE, TESTING_GUIDE, DEVELOPER_GUIDE)
-# 0.9.1 (2025-11-10):
-#   - Fix WAV audio setting not persisting when saved
-# 0.9.0 (2025-11-10):
-#   - Add iOS Safari audio compatibility (WAV conversion)
-#   - Fix audio generation deadlock with subprocess.DEVNULL
-#   - Add user-adjustable silence trimming threshold
-#   - Improve CCS testing framework integration
-#   - Implement edit distance scoring algorithm
-#   - Add version management system
+# Translation module
+from translation import (
+    get_translation_provider, validate_translation_api_key,
+    get_translation_from_llm, enrich_material_file,
+)
 
 import streamlit as st
 
@@ -221,402 +106,29 @@ except ImportError:
 
 
 # ============================================================================
-# SETTINGS FUNCTIONS (must be defined before authentication)
+# SETTINGS FUNCTIONS — thin wrappers over config.py
 # ============================================================================
 
 def load_settings():
-    """Load user settings from database (if authenticated) or local config file"""
-    default_settings = {
-        "speed": 140,
-        "pitch": 35,
-        "voice": "pt-br",
-        "model": "base",
-        "duration": 3,
-        "comparison_algorithm": "edit_distance",  # or "positional"
-        "asr_engine": "whisper",  # "whisper" or "wav2vec2"
-        "whisper_model_size": "base",  # tiny, base, small, medium, large
-        "silence_threshold": 0.01,  # Energy threshold for silence detection (0.001-0.1)
-        "use_wav_audio": False,  # Convert TTS audio to WAV for iOS Safari compatibility
-        "tts_engine": "google_cloud",  # "google_cloud" (official API, best), "gtts" (unofficial, rate limited), or "espeak"
-        "gtts_slow": False,  # Enable slow speech for Google TTS (when tts_engine='gtts')
-    }
-
-    # If authenticated, load from database
-    if st.session_state.get('authenticated', False) and 'user' in st.session_state:
-        try:
-            user_id = st.session_state['user']['user_id']
-            db_settings = app_mysql.get_user_settings(user_id)
-            if db_settings:
-                default_settings.update(db_settings)
-                return default_settings
-        except Exception as e:
-            st.warning(f"Could not load settings from database: {e}")
-
-    # Fall back to local config file for non-authenticated users
-    config_file = Path("practice_config.json")
-    if config_file.exists():
-        try:
-            with open(config_file) as f:
-                saved = json.load(f)
-                default_settings.update(saved)
-        except Exception:
-            pass
-
-    return default_settings
+    """Load user settings from database or local config file."""
+    from config import load_settings as _load
+    return _load(session_state=st.session_state, db_module=app_mysql)
 
 
 def save_settings(settings: Dict):
-    """Save settings to database (if authenticated) or local config file"""
-    # If authenticated, save to database
-    if st.session_state.get('authenticated', False) and 'user' in st.session_state:
-        try:
-            user_id = st.session_state['user']['user_id']
-            # Save each setting individually to database
-            for key, value in settings.items():
-                app_mysql.save_user_setting(user_id, key, value)
-            return
-        except Exception as e:
-            st.error(f"Could not save settings to database: {e}")
-            return
-
-    # Fall back to local config file for non-authenticated users
-    try:
-        with open("practice_config.json", 'w') as f:
-            json.dump(settings, f, indent=2)
-    except Exception as e:
-        st.error(f"Could not save settings: {e}")
+    """Save settings to database or local config file."""
+    from config import save_settings as _save
+    _save(settings, session_state=st.session_state, db_module=app_mysql,
+          error_callback=st.error)
 
 
 # ============================================================================
-# MATERIAL ENRICHMENT (LLM Translations + IPA Generation)
+# MATERIAL ENRICHMENT — now in translation.py (imported above)
+# The imported functions accept secrets/db_module as parameters.
+# Call sites in this file that previously used st.secrets / app_mysql directly
+# now go through the imported versions. Where needed, pass st.secrets and
+# app_mysql explicitly.
 # ============================================================================
-
-def get_translation_provider() -> str:
-    return (
-        st.secrets.get("translation_provider")
-        or os.environ.get("TRANSLATION_PROVIDER")
-        or "google"
-    ).lower()
-
-
-def get_language_code(language_name: str) -> str:
-    if language_name.lower() == "english":
-        return "en"
-    if language_name in LANGUAGE_CONFIG:
-        return LANGUAGE_CONFIG[language_name].get("code", language_name.lower())
-    return language_name.lower()
-
-
-def get_language_for_provider(provider: str, language_name: str) -> str:
-    # Google expects language codes; OpenAI can use full names
-    if provider == "google":
-        return get_language_code(language_name)
-    return language_name
-
-
-def validate_translation_api_key(provider: str) -> tuple[bool, str]:
-    """
-    Validate translation API key for the selected provider.
-    """
-    provider = (provider or "google").lower()
-
-    if provider == "google":
-        api_key = st.secrets.get("google_cloud_translate_api_key") or os.environ.get("GOOGLE_TRANSLATE_API_KEY")
-        if not api_key or api_key == "your-google-translate-api-key-here":
-            return False, "Valid Google Translate API key required. Configure google_cloud_translate_api_key in secrets.toml or set GOOGLE_TRANSLATE_API_KEY."
-        return True, api_key
-
-    if provider == "openai":
-        api_key = st.secrets.get("openai_api_key") or os.environ.get("OPENAI_API_KEY")
-        if not api_key or api_key == "your-openai-api-key-here":
-            return False, "Valid OpenAI API key required. Configure openai_api_key in secrets.toml or set OPENAI_API_KEY."
-        return True, api_key
-
-    return False, f"Unknown translation provider: {provider}"
-
-
-def get_ipa_from_espeak(text: str, lang_code: str) -> str:
-    """
-    Generate IPA transcription using espeak-ng.
-
-    Args:
-        text: Text to transcribe
-        lang_code: Language code (pt, fr, nl, de, it, es)
-
-    Returns:
-        IPA transcription or '[error]' on failure
-    """
-    # Map language codes to espeak voices
-    ESPEAK_LANG_MAP = {
-        'pt': 'pt-br',
-        'fr': 'fr-fr',
-        'nl': 'nl',
-        'de': 'de',
-        'it': 'it',
-        'es': 'es'
-    }
-
-    espeak_lang = ESPEAK_LANG_MAP.get(lang_code, lang_code)
-    espeak_cmd = get_espeak_path()
-
-    try:
-        result = subprocess.run(
-            [espeak_cmd, '-v', espeak_lang, '-q', '--ipa', text],
-            capture_output=True,
-            text=True,
-            timeout=5
-        )
-        if result.returncode == 0:
-            ipa = result.stdout.strip()
-            # Clean up spacing
-            ipa = ' '.join(ipa.split())
-            return ipa
-        return '[error]'
-    except subprocess.TimeoutExpired:
-        return '[timeout]'
-    except Exception as e:
-        return f'[error: {str(e)}]'
-
-
-def get_translation_from_llm(text: str, source_lang: str, target_lang: str = "English") -> str:
-    """
-    Get translation using a pluggable provider (default: Google Translate).
-    """
-    try:
-        provider = get_translation_provider()
-
-        # Validate API key
-        is_valid, api_key_or_error = validate_translation_api_key(provider)
-        if not is_valid:
-            return f"[error: {api_key_or_error}]"
-
-        from translation_providers import get_translator
-
-        source_lang_for_provider = get_language_for_provider(provider, source_lang)
-        target_lang_for_provider = get_language_for_provider(provider, target_lang)
-
-        # Cache lookup
-        cached = app_mysql.get_translation_cache(
-            source_lang=source_lang_for_provider,
-            target_lang=target_lang_for_provider,
-            source_text=text,
-            provider=provider,
-        )
-        if cached and cached.get("translated_text"):
-            return cached["translated_text"]
-
-        translator = get_translator(provider, api_key=api_key_or_error)
-        result = translator.translate(text, source_lang_for_provider, target_lang_for_provider)
-
-        # Cache store
-        app_mysql.set_translation_cache(
-            source_lang=source_lang_for_provider,
-            target_lang=target_lang_for_provider,
-            source_text=text,
-            translated_text=result.translated_text,
-            provider=provider,
-            detected_source=result.detected_source,
-            confidence=result.confidence,
-        )
-
-        # Log API usage for cost tracking
-        try:
-            if provider == "openai" and getattr(result.raw, "usage", None):
-                log_api_call(
-                    api_name='openai',
-                    model=getattr(result.raw, "model", "gpt-4o-mini"),
-                    operation='translation',
-                    input_tokens=result.raw.usage.prompt_tokens,
-                    output_tokens=result.raw.usage.completion_tokens,
-                    metadata={'source_lang': source_lang, 'target_lang': target_lang}
-                )
-            else:
-                log_api_call(
-                    api_name=provider,
-                    model='google-translate',
-                    operation='translation',
-                    input_tokens=0,
-                    output_tokens=0,
-                    metadata={'source_lang': source_lang, 'target_lang': target_lang}
-                )
-        except Exception:
-            pass
-
-        return result.translated_text
-
-    except Exception as e:
-        return f"[error: {str(e)}]"
-
-
-def enrich_material_file(
-    file_path: Path,
-    lang_code: str,
-    add_translations: bool = True,
-    add_ipa: bool = True,
-    progress_callback=None
-) -> Dict:
-    """
-    Enrich a material file by adding missing translations and/or IPA.
-
-    Args:
-        file_path: Path to the material file
-        lang_code: Language code (pt, fr, nl, etc.)
-        add_translations: Whether to add missing translations
-        add_ipa: Whether to add missing IPA
-        progress_callback: Optional callback function(current, total, message)
-
-    Returns:
-        Dict with keys: success (bool), message (str), stats (dict)
-    """
-    provider = get_translation_provider()
-
-    if add_translations:
-        # Validate API key before proceeding
-        is_valid, error_message = validate_translation_api_key(provider)
-        if not is_valid:
-            return {
-                'success': False,
-                'message': error_message,
-                'stats': {}
-            }
-
-    # Map language codes to full names for LLM
-    LANG_NAMES = {
-        'pt': 'Portuguese',
-        'fr': 'French',
-        'nl': 'Dutch',
-        'de': 'German',
-        'it': 'Italian',
-        'es': 'Spanish'
-    }
-    source_lang_name = LANG_NAMES.get(lang_code, lang_code.upper())
-
-    # Read file
-    try:
-        with open(file_path, 'r', encoding='utf-8') as f:
-            lines = f.readlines()
-    except Exception as e:
-        return {
-            'success': False,
-            'message': f'Could not read file: {e}',
-            'stats': {}
-        }
-
-    # Create backup
-    backup_path = file_path.with_suffix('.bak')
-    try:
-        with open(backup_path, 'w', encoding='utf-8') as f:
-            f.writelines(lines)
-    except Exception as e:
-        return {
-            'success': False,
-            'message': f'Could not create backup: {e}',
-            'stats': {}
-        }
-
-    # Process lines
-    enriched_lines = []
-    stats = {
-        'total_lines': 0,
-        'translations_added': 0,
-        'ipa_added': 0,
-        'errors': []
-    }
-
-    for i, line in enumerate(lines):
-        # Skip comments and empty lines
-        if line.strip().startswith('#') or not line.strip():
-            enriched_lines.append(line)
-            continue
-
-        # Normalize: strip trailing pipes and whitespace
-        normalized_line = line.strip()
-        while normalized_line.endswith('|'):
-            normalized_line = normalized_line[:-1].strip()
-
-        # Parse the line by splitting on pipes
-        if '|' in normalized_line:
-            parts = [p.strip() for p in normalized_line.split('|')]
-            phrase = parts[0] if len(parts) > 0 else ''
-            translation = parts[1] if len(parts) > 1 else ''
-            ipa = parts[2] if len(parts) > 2 else ''
-        else:
-            # Plain text format (no pipes) - treat entire line as phrase
-            phrase = normalized_line
-            translation = ''
-            ipa = ''
-
-        # Skip if no phrase
-        if not phrase:
-            enriched_lines.append(line)
-            continue
-
-        stats['total_lines'] += 1
-
-        # Update progress
-        if progress_callback:
-            progress_callback(i + 1, len(lines), f"Processing: {phrase[:30]}...")
-
-        # Add translation if missing
-        if add_translations and not translation:
-            # Debug: log what we're sending to LLM
-            if progress_callback:
-                progress_callback(i + 1, len(lines), f"Translating: {phrase[:30]}...")
-
-            new_translation = get_translation_from_llm(phrase, source_lang_name)
-
-            # Debug: check if LLM returned IPA instead of translation
-            if new_translation.startswith('[') and not new_translation.startswith('[error'):
-                stats['errors'].append(f"LLM returned IPA instead of translation for '{phrase}': {new_translation}")
-            elif not new_translation.startswith('[error'):
-                translation = new_translation
-                stats['translations_added'] += 1
-            else:
-                stats['errors'].append(f"Translation error for '{phrase}': {new_translation}")
-
-        # Add IPA if missing
-        # Consider IPA missing if empty or just placeholder markers
-        ipa_empty = not ipa or ipa in ['[ipa]', '[]']
-        if add_ipa and ipa_empty:
-            new_ipa = get_ipa_from_espeak(phrase, lang_code)
-            if not new_ipa.startswith('[error') and not new_ipa.startswith('[timeout') and new_ipa.strip():
-                ipa = f"[{new_ipa}]"  # Wrap in brackets
-                stats['ipa_added'] += 1
-            else:
-                if new_ipa.strip():
-                    stats['errors'].append(f"IPA error for '{phrase}': {new_ipa}")
-                else:
-                    stats['errors'].append(f"IPA empty for '{phrase}'")
-
-        # Reconstruct line with consistent format: always 3 fields
-        enriched_line = f"{phrase} | {translation} | {ipa}\n"
-        enriched_lines.append(enriched_line)
-
-    # Write enriched content
-    try:
-        with open(file_path, 'w', encoding='utf-8') as f:
-            f.writelines(enriched_lines)
-    except Exception as e:
-        # Restore from backup
-        try:
-            with open(backup_path, 'r', encoding='utf-8') as f:
-                backup_content = f.read()
-            with open(file_path, 'w', encoding='utf-8') as f:
-                f.write(backup_content)
-        except Exception:
-            pass
-
-        return {
-            'success': False,
-            'message': f'Could not write enriched file: {e}',
-            'stats': stats
-        }
-
-    return {
-        'success': True,
-        'message': 'File enriched successfully',
-        'stats': stats
-    }
 
 
 # ============================================================================
@@ -1272,120 +784,9 @@ def get_wav2vec2_model():
     return st.session_state.wav2vec2_processor, st.session_state.wav2vec2_model
 
 
-def get_espeak_path():
-    """
-    Get espeak path (local build or system-wide)
-
-    Platform differences:
-    - macOS (MacPorts): Binary is "espeak" at /opt/local/bin/espeak
-    - Debian/Ubuntu (Streamlit Cloud): Binary is "espeak-ng" from espeak-ng package
-    """
-    # Try macOS MacPorts path first
-    local_path = "/opt/local/bin/espeak"
-    if Path(local_path).exists():
-        return local_path
-
-    # Try espeak-ng (Streamlit Cloud / Ubuntu)
-    try:
-        subprocess.run(["espeak-ng", "--version"], capture_output=True, check=True)
-        return "espeak-ng"
-    except (subprocess.CalledProcessError, FileNotFoundError):
-        pass
-
-    # Fallback to espeak (if available)
-    return "espeak"
-
-
-def get_phonemes(text: str, voice: str = "pt-br") -> str:
-    """Get eSpeak phoneme codes (eIPA) for text"""
-    try:
-        espeak_cmd = get_espeak_path()
-        result = subprocess.run(
-            [espeak_cmd, "-v", voice, "-x", "-q", text],
-            capture_output=True,
-            text=True,
-            check=True
-        )
-        # eSpeak may insert newlines around punctuation; normalize to single spaces
-        phonemes = result.stdout.strip()
-        phonemes = ' '.join(phonemes.split())
-        return phonemes
-    except (subprocess.CalledProcessError, FileNotFoundError) as e:
-        st.warning(f"eSpeak phoneme error: {e} (cmd: {get_espeak_path()})")
-        return "[phonemes unavailable]"
-
-
-def normalize_for_phoneme_scoring(s: str) -> str:
-    """
-    Normalize eSpeak phoneme strings for pronunciation scoring.
-
-    Removes:
-    - All whitespace (word boundaries, spaces)
-    - Pause phonemes (_: _! _| _:: etc.) - inserted by eSpeak for punctuation
-
-    This ensures scoring is based purely on pronunciation phonemes,
-    not on text formatting artifacts like quotes, commas, periods.
-    """
-    import re
-    if not s:
-        return ""
-    # Remove all whitespace
-    s = re.sub(r"\s+", "", s.strip())
-    # Remove eSpeak pause phonemes: _: _! _| _:: and combinations like _:_:
-    # These are inserted for punctuation and should not affect pronunciation scoring
-    s = re.sub(r'_[:!|]+', '', s)
-    return s
-
-
-def get_ipa(text: str, voice: str = "pt-br") -> str:
-    """
-    Get IPA transcription for text
-
-    Note: eSpeak converts punctuation (commas, periods, etc.) into newlines
-    in the IPA output. We normalize these to spaces for consistent comparison.
-    """
-    try:
-        espeak_cmd = get_espeak_path()
-        result = subprocess.run(
-            [espeak_cmd, "-v", voice, "--ipa", "-q", text],
-            capture_output=True,
-            text=True,
-            check=True
-        )
-        # Strip leading/trailing whitespace and normalize internal newlines to spaces
-        # eSpeak inserts newlines at punctuation (commas, periods, etc.)
-        ipa = result.stdout.strip()
-        ipa = ' '.join(ipa.split())  # Replace all whitespace (including \n) with single space
-        return ipa
-    except (subprocess.CalledProcessError, FileNotFoundError) as e:
-        st.warning(f"eSpeak IPA error: {e} (cmd: {get_espeak_path()})")
-        return "[IPA unavailable]"
-
-
-def format_ipa(ipa_text: str, size: str = "1.0em", weight: int = 400, brackets: bool = True) -> str:
-    """
-    Format IPA text with consistent delimiters and styling.
-
-    Args:
-        ipa_text: The IPA transcription text
-        size: Font size (default "1.0em" for standard, "0.9em" for smaller)
-        weight: Font weight (default 400 for normal, 300 for light)
-        brackets: Whether to include square brackets (default True)
-
-    Returns HTML-formatted IPA with consistent styling.
-    Example: <span style="font-size: 1.0em; font-weight: 400;">[ipˈa]</span>
-    """
-    if not ipa_text:
-        return ""
-
-    # Remove any existing delimiters
-    ipa_clean = ipa_text.strip().strip('[]/()')
-
-    # Add brackets if requested
-    display_text = f"[{ipa_clean}]" if brackets else ipa_clean
-
-    # Return formatted HTML
-    return f'<span style="font-size: {size}; font-weight: {weight}; font-family: \'Doulos SIL\', \'Charis SIL\', \'Gentium Plus\', \'DejaVu Sans\', sans-serif;">{display_text}</span>'
+# Phoneme/IPA functions — now in scoring/phonemes.py (imported above)
+# espeak path, get_phonemes, normalize_for_phoneme_scoring, get_ipa,
+# get_ipa_from_espeak, format_ipa are all imported at top of file.
 
 
 def speak_text(text: str, voice: str = "pt-br", speed: int = 160, pitch: int = 40) -> tuple[bytes, str]:
@@ -1800,144 +1201,9 @@ def transcribe_audio(audio_file: str, settings: Dict, language: str = "Portugues
     return transcribe_audio_whisper(audio_file, model, lang_code)
 
 
-def levenshtein_distance(s1: str, s2: str) -> int:
-    """
-    Calculate Levenshtein (edit) distance between two strings.
-    Returns the minimum number of single-character edits (insertions, deletions, substitutions).
-    """
-    if len(s1) < len(s2):
-        return levenshtein_distance(s2, s1)
-
-    if len(s2) == 0:
-        return len(s1)
-
-    previous_row = range(len(s2) + 1)
-    for i, c1 in enumerate(s1):
-        current_row = [i + 1]
-        for j, c2 in enumerate(s2):
-            # Cost of insertions, deletions, or substitutions
-            insertions = previous_row[j + 1] + 1
-            deletions = current_row[j] + 1
-            substitutions = previous_row[j] + (c1 != c2)
-            current_row.append(min(insertions, deletions, substitutions))
-        previous_row = current_row
-
-    return previous_row[-1]
-
-
-def get_edit_operations(s1: str, s2: str):
-    """
-    Get the actual edit operations needed to transform s1 into s2.
-    Returns a list of tuples: (operation, position, char1, char2)
-    where operation is 'match', 'substitute', 'insert', or 'delete'
-    """
-    # Build the DP table
-    m, n = len(s1), len(s2)
-    dp = [[0] * (n + 1) for _ in range(m + 1)]
-
-    for i in range(m + 1):
-        dp[i][0] = i
-    for j in range(n + 1):
-        dp[0][j] = j
-
-    for i in range(1, m + 1):
-        for j in range(1, n + 1):
-            if s1[i-1] == s2[j-1]:
-                dp[i][j] = dp[i-1][j-1]
-            else:
-                dp[i][j] = 1 + min(
-                    dp[i-1][j],    # delete
-                    dp[i][j-1],    # insert
-                    dp[i-1][j-1]   # substitute
-                )
-
-    # Backtrack to find operations
-    operations = []
-    i, j = m, n
-    while i > 0 or j > 0:
-        if i > 0 and j > 0 and s1[i-1] == s2[j-1]:
-            operations.append(('match', i-1, s1[i-1], s2[j-1]))
-            i -= 1
-            j -= 1
-        elif i > 0 and j > 0 and dp[i][j] == dp[i-1][j-1] + 1:
-            operations.append(('substitute', i-1, s1[i-1], s2[j-1]))
-            i -= 1
-            j -= 1
-        elif j > 0 and dp[i][j] == dp[i][j-1] + 1:
-            operations.append(('insert', i, '-', s2[j-1]))
-            j -= 1
-        elif i > 0 and dp[i][j] == dp[i-1][j] + 1:
-            operations.append(('delete', i-1, s1[i-1], '-'))
-            i -= 1
-
-    operations.reverse()
-    return operations
-
-
-def compare_phonemes_positional(user_phonemes: str, correct_phonemes: str):
-    """
-    DEPRECATED: Simple positional matching (fails with insertions/deletions).
-    Kept for reference but not recommended for use.
-    """
-    exact_match = user_phonemes == correct_phonemes
-
-    if len(correct_phonemes) == 0:
-        similarity = 0.0
-    else:
-        matches = sum(
-            1 for a, b in zip(user_phonemes, correct_phonemes)
-            if a == b
-        )
-        similarity = matches / max(len(user_phonemes), len(correct_phonemes))
-
-    return exact_match, similarity
-
-
-def compare_phonemes_edit_distance(user_phonemes: str, correct_phonemes: str):
-    """
-    Compare phonemes using edit distance (Levenshtein).
-    Handles insertions, deletions, and substitutions gracefully.
-
-    Returns:
-        exact_match: bool - True if strings are identical
-        similarity: float - 0.0 to 1.0, where 1.0 is perfect match
-        distance: int - Number of edits needed
-    """
-    exact_match = user_phonemes == correct_phonemes
-
-    if len(correct_phonemes) == 0:
-        return exact_match, 0.0, len(user_phonemes)
-
-    distance = levenshtein_distance(user_phonemes, correct_phonemes)
-    max_length = max(len(user_phonemes), len(correct_phonemes))
-
-    # Similarity: 1.0 means perfect match, 0.0 means completely different
-    similarity = 1.0 - (distance / max_length)
-
-    return exact_match, similarity, distance
-
-
-def compare_phonemes(user_phonemes: str, correct_phonemes: str, algorithm: str = "edit_distance"):
-    """
-    Modular phoneme comparison with selectable algorithms.
-
-    Args:
-        user_phonemes: Phonemes from user's speech
-        correct_phonemes: Target phonemes
-        algorithm: "edit_distance" (default) or "positional"
-
-    Returns:
-        exact_match: bool
-        similarity: float (0.0 to 1.0)
-        distance: int (only for edit_distance, None otherwise)
-    """
-    if algorithm == "edit_distance":
-        return compare_phonemes_edit_distance(user_phonemes, correct_phonemes)
-    elif algorithm == "positional":
-        exact_match, similarity = compare_phonemes_positional(user_phonemes, correct_phonemes)
-        return exact_match, similarity, None
-    else:
-        raise ValueError(f"Unknown algorithm: {algorithm}")
+# Scoring/comparison functions — now in scoring/comparison.py (imported above)
+# levenshtein_distance, get_edit_operations, compare_phonemes_positional,
+# compare_phonemes_edit_distance, compare_phonemes are all imported at top.
 
 
 def practice_word_from_audio(text: str, audio_bytes: bytes, settings: Dict):
