+++ tests/test_config.py
@@ -1,7 +1,7 @@
 """
 Tests for configuration and language utilities.
 
-These test pure data lookups from LANGUAGE_CONFIG — no Streamlit needed.
+Now imports from the extracted config module directly.
 """
 
 import sys
@@ -9,27 +9,7 @@ from pathlib import Path
 
 sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
 
-# LANGUAGE_CONFIG is defined at module level in app.py but requires Streamlit
-# to fully import. We replicate the lookup functions here as a baseline.
-# Post-refactor, these import from config.py directly.
-
-LANGUAGE_CONFIG = {
-    "Portuguese": {"code": "pt"},
-    "Dutch": {"code": "nl"},
-    "French": {"code": "fr"},
-    "German": {"code": "de"},
-    "Spanish": {"code": "es"},
-    "Italian": {"code": "it"},
-}
-
-
-def get_language_code(language_name: str) -> str:
-    """Copied from app.py:302"""
-    if language_name.lower() == "english":
-        return "en"
-    if language_name in LANGUAGE_CONFIG:
-        return LANGUAGE_CONFIG[language_name].get("code", language_name.lower())
-    return language_name.lower()
+from config import LANGUAGE_CONFIG, get_language_code, DEFAULT_SETTINGS
 
 
 class TestGetLanguageCode:
@@ -44,3 +24,22 @@ class TestGetLanguageCode:
 
     def test_unknown_falls_back_to_lowercase(self):
         assert get_language_code("Swahili") == "swahili"
+
+
+class TestLanguageConfig:
+    def test_all_languages_have_code(self):
+        for lang, cfg in LANGUAGE_CONFIG.items():
+            assert "code" in cfg, f"{lang} missing 'code'"
+
+    def test_all_languages_have_voices(self):
+        for lang, cfg in LANGUAGE_CONFIG.items():
+            assert "voices" in cfg, f"{lang} missing 'voices'"
+            for engine in ("google_cloud", "gtts", "espeak"):
+                assert engine in cfg["voices"], f"{lang} missing voice for {engine}"
+
+
+class TestDefaultSettings:
+    def test_has_required_keys(self):
+        required = ["tts_engine", "asr_engine", "comparison_algorithm", "voice"]
+        for key in required:
+            assert key in DEFAULT_SETTINGS, f"Missing default setting: {key}"
